refactor(step2): 리뷰 바탕으로 리팩토링
