+++ src/main/kotlin/study/step2/ExpressionFactory.kt
@@ -1,0 +1,11 @@
+package study.step2
+
+class ExpressionFactory(
+    private val tokenizer: Tokenizer
+) {
+    fun create(expressionStr: String?): Expression {
+        require(!expressionStr.isNullOrBlank())
+        val rawTokens = tokenizer.tokenize(expressionStr)
+        return Expression(rawTokens)
+    }
+}
