+++ src/main/kotlin/study/step2/Tokenizer.kt
@@ -1,0 +1,11 @@
+package study.step2
+
+class Tokenizer(
+    private val tokenListValidator: TokenListValidator
+) {
+    fun tokenize(expressionStr: String): List<String> {
+        val rawTokens = expressionStr.split(" ")
+        tokenListValidator.validate(rawTokens)
+        return rawTokens
+    }
+}
